Extract shared order and refund methods to trait
