+++ includes/class-wc-admin-api-init.php
@@ -43,6 +43,7 @@ class WC_Admin_Api_Init {
 		require_once WC_ADMIN_ABSPATH . 'includes/class-wc-admin-reports-parameter-exception.php';
 
 		// WC Class extensions.
+		require_once WC_ADMIN_ABSPATH . 'includes/class-wc-admin-order-trait.php';
 		require_once WC_ADMIN_ABSPATH . 'includes/class-wc-admin-order.php';
 		require_once WC_ADMIN_ABSPATH . 'includes/class-wc-admin-order-refund.php';
